Move connection configuration classes to the "connection" package
