+++ publishing-example-java-app/src/test/java/com/ably/tracking/example/javapublisher/PublisherInterfaceUsageExamples.java
@@ -6,9 +6,9 @@ import android.content.Context;
 import com.ably.tracking.Accuracy;
 import com.ably.tracking.BuilderConfigurationIncompleteException;
 import com.ably.tracking.ConnectionException;
-import com.ably.tracking.ConnectionConfiguration;
+import com.ably.tracking.connection.ConnectionConfiguration;
 import com.ably.tracking.Resolution;
-import com.ably.tracking.TokenRequest;
+import com.ably.tracking.connection.TokenRequest;
 import com.ably.tracking.java.ConnectionConfigurationFactory;
 import com.ably.tracking.publisher.DefaultProximity;
 import com.ably.tracking.publisher.DefaultResolutionConstraints;
