+++ publishing-sdk/src/main/java/com/ably/tracking/publisher/PublisherBuilder.kt
@@ -5,7 +5,7 @@ import android.Manifest.permission.ACCESS_FINE_LOCATION
 import android.content.Context
 import androidx.annotation.RequiresPermission
 import com.ably.tracking.BuilderConfigurationIncompleteException
-import com.ably.tracking.ConnectionConfiguration
+import com.ably.tracking.connection.ConnectionConfiguration
 import com.ably.tracking.common.DefaultAbly
 
 internal data class PublisherBuilder(
